fsdfdfdfdsfd

+++ ToolKids/ToolKids/Device_Info.swift
@@ -21,6 +21,6 @@ class Device_Info : AnyObject {
     var area_code : String?
     var imsi : String?
     var iccid : String?
-    var sn : Int8?
+    var sn : Int64?
     var rnd : String?
 }
